标签管理接上接口。

+++ label/catalog_label.py
@@ -10,6 +10,8 @@ from product import models as product_models
 
 import models
 from util import sync_util
+from panda.settings import PRODUCT_POOL_OWNER_ID
+
 
 #分类配置标签
 class CataloLabel(resource.Resource):
@@ -69,15 +71,18 @@ class CataloLabel(resource.Resource):
 		select_catalog_labels = request.POST.get('select_catalog_labels', '')
 		catalog_id = request.POST.get('catalog_id', -1)
 		product_id = int(request.POST.get('product_id', -1))
+
 		if select_catalog_labels:
 			select_catalog_labels = json.loads(select_catalog_labels)
 			
 			catalog_label_create = []
 			product_label_create = []
 			property_id_and_value_ids = []
+			label_ids = []
 			for select_catalog_label in select_catalog_labels:
 				value_ids = []
 				select_value_ids = select_catalog_label['valueIds']
+				label_ids += select_value_ids
 				property_id = select_catalog_label['propertyId']
 				for value_id in select_value_ids:
 					value_ids.append(str(value_id))
@@ -99,13 +104,47 @@ class CataloLabel(resource.Resource):
 				))
 
 			if product_id == -1:
-				#分类关联标签
-				catalog_models.ProductCatalogHasLabel.objects.filter(catalog_id=catalog_id).delete()
-				catalog_models.ProductCatalogHasLabel.objects.bulk_create(catalog_label_create)
-			else:
-				#商品关联标签
-				product_models.ProductHasLabel.objects.filter(product_id=product_id).delete()
-				product_models.ProductHasLabel.objects.bulk_create(product_label_create)
 
+				# 类目下的标签
+				catalog_relation = catalog_models.ProductCatalogRelation.objects.filter(catalog_id=catalog_id).first()
+				# label_ids = []
+				# label_ids = select_catalog_labels[0].get('valueIds')
+				label_relations = models.LabelGroupValueRelation.objects.filter(label_value_id__in=label_ids)
+
+				weapp_classification_id = catalog_relation.weapp_catalog_id
+				params = {
+					'label_ids': json.dumps([relation.weapp_label_value_id for relation in label_relations]),
+					'classification_id': weapp_classification_id,
+					'owner_id': PRODUCT_POOL_OWNER_ID
+				}
+				resp, resp_data = sync_util.sync_zeus(params=params, resource='mall.classification_has_label', method='post')
+				if resp:
+					# 分类关联标签
+					catalog_models.ProductCatalogHasLabel.objects.filter(catalog_id=catalog_id).delete()
+					catalog_models.ProductCatalogHasLabel.objects.bulk_create(catalog_label_create)
+				else:
+					response = create_response(500)
+					return response.get_response()
+			else:
+
+				product_relation = product_models.ProductHasRelationWeapp.objects.filter(product_id=product_id).first()
+				weapp_product_id = product_relation.weapp_product_id
+
+				# label_ids = select_catalog_labels[0].get('valueIds')
+				label_relations = models.LabelGroupValueRelation.objects.filter(label_value_id__in=label_ids)
+				params = {
+					'product_id': weapp_product_id,
+					'owner_id': PRODUCT_POOL_OWNER_ID,
+					'label_ids': json.dumps([relation.weapp_label_value_id for relation in label_relations]),
+				}
+				resp, resp_data = sync_util.sync_zeus(params=params, resource='mall.product_has_label',
+													  method='post')
+				if resp:
+					# 商品关联标签
+					product_models.ProductHasLabel.objects.filter(product_id=product_id).delete()
+					product_models.ProductHasLabel.objects.bulk_create(product_label_create)
+				else:
+					response = create_response(500)
+					return response.get_response()
 		response = create_response(200)
 		return response.get_response()
